requirements.txt
Flask==2.2.3
mitmproxy==7.0.4
requests==2.28.2
tldextract==3.1.0
pytest
python-magic-bin
